feat: 멀티모델 LiDAR 지원 (LDROBOT STL-27L + Pacecat LDS-50C-E)
reader를 공용 계층 + 제조사/모델별 코덱으로 재구성해 뷰어 하나로 여러
모델을 쓸 수 있게 함. 기존 파이프라인(raw→parse→frame→viewer) 유지.

- 추상화: Decoder 트레잇(model.rs) + Transport 트레잇(transport.rs)으로
  모델·전송(시리얼/UDP)을 가림. data_collector는 모델/전송 독립 + start/stop 제어.
- Pacecat LDS-50C-E(UDP 0xFAC7): frame 파싱(28B 헤더 + dist/ang/strength SoA +
  16-bit sum), 스트리밍 디코더, LVERSH/LSTARH 시작 커맨드(STM32 CRC32), UDP 연결.
  실하드웨어로 검증(CRC 불일치 0, ~1700pts/회전).
- STL-27L 코덱을 reader/LDROBOT/STL-27L/ 로 이동(폴더는 #[path]로 연결).
- 뷰어: 라이다 개수·모델 드롭다운, Start/Stop 버튼(누를 때 연결 시작),
  설정 패널 토글, 장치별 색 구분 렌더.
- config.toml: model + UDP 네트워크 필드 추가.
- docs/ARCHITECTURE.md 멀티모델 구조로 갱신.

+++ src/reader/LDROBOT/STL-27L/frame.rs
@@ -6,7 +6,7 @@
 use zerocopy::byteorder::little_endian::U16;
 use zerocopy::{FromBytes, Immutable, KnownLayout, Unaligned};
 
-use super::types::{LidarBody, LidarPoint, ParseError};
+use crate::reader::types::{LidarBody, LidarPoint, ParseError};
 
 pub const HEADER: u8 = 0x54;
 pub const VER_LEN: u8 = 0x2c;
@@ -85,7 +85,7 @@ pub fn parse(data: &[u8]) -> Result<LidarBody, ParseError> {
     })?;
 
     if raw.header != HEADER {
-        return Err(ParseError::InvalidHeader(raw.header));
+        return Err(ParseError::InvalidHeader(raw.header as u16));
     }
 
     if raw.body.ver_len != VER_LEN {
@@ -95,8 +95,8 @@ pub fn parse(data: &[u8]) -> Result<LidarBody, ParseError> {
     let expected_crc = crc8(&data[..CRC_OFFSET]);
     if expected_crc != raw.crc8 {
         return Err(ParseError::CrcMismatch {
-            expected: expected_crc,
-            actual: raw.crc8,
+            expected: expected_crc as u32,
+            actual: raw.crc8 as u32,
         });
     }
 
@@ -107,7 +107,7 @@ pub fn parse(data: &[u8]) -> Result<LidarBody, ParseError> {
         speed_degrees_per_second: raw.body.speed.get(),
         start_angle_degrees: raw_angle_to_degrees(start_angle_raw as f32),
         end_angle_degrees: raw_angle_to_degrees(end_angle_raw as f32),
-        timestamp_ms: raw.body.timestamp.get(),
+        timestamp_ms: raw.body.timestamp.get() as u32,
         points: build_points(&raw.body.points, start_angle_raw, end_angle_raw),
     })
 }
